+++ src/main.rs
@@ -1,18 +1,24 @@
 #![no_std] // disable the Rust standard library
 #![no_main] // disable the Rust entry point
-#![feature(custom_test_frameworks)] // enable custom test framework
-#![test_runner(crate::test_runner)] // point to our custom runner
+#![feature(custom_test_frameworks)]
+#![test_runner(rost::test_runner)]
 #![reexport_test_harness_main = "test_main"]
 
-extern crate rlibc;
-
-mod vga_buffer;
-mod serial;
-
 use core::panic::PanicInfo;
+use rost::println;
+
+/// System entry point.
+/// Must be named _start by convention.
+#[no_mangle] // output non-cryptic function name
+pub extern "C" fn _start() -> ! {
+    println!("Hello world!");
+    #[cfg(test)]
+    test_main();
+    loop {}
+}
 
 
-/// Automatically called on panic.
+/// Called on panic.
 #[cfg(not(test))]
 #[panic_handler]
 fn panic(info: &PanicInfo) -> ! {
@@ -23,30 +29,7 @@ fn panic(info: &PanicInfo) -> ! {
 #[cfg(test)]
 #[panic_handler]
 fn panic(info: &PanicInfo) -> ! {
-    serial_println!("[failed]\n");
-    serial_println!("Error: {}\n", info);
-    exit_qemu(QemuExitCode::Failure);
-    loop {}
-}
-/// System entry point.
-/// Must be named _start by convention.
-#[no_mangle] // output non-cryptic function name
-pub extern "C" fn _start() -> ! {
-    println!("Hello world!");
-
-    #[cfg(test)]
-    test_main();
-
-    loop {}
-}
-
-#[cfg(test)]
-pub fn test_runner(tests: &[&dyn Testable]) {
-    serial_println!("Running {} tests", tests.len());
-    for test in tests {
-        test.run();
-    }
-    exit_qemu(QemuExitCode::Success);
+    rost::test_panic_handler(info);
 }
 
 #[test_case]
@@ -54,35 +37,4 @@ fn trivial_test() {
     assert_eq!(1, 1);
 }
 
-#[derive(Debug, Clone, PartialEq, Eq)]
-#[repr(u32)]
-pub enum QemuExitCode {
-    Success = 0x10,
-    Failure = 0x11,
-}
 
-/// Tell QEMU to exit.
-pub fn exit_qemu(exit_code: QemuExitCode) {
-    use x86_64::instructions::port::Port;
-    unsafe {
-        // port offset and size defined in `Cargo.toml`
-        let mut port = Port::new(0xf4);
-        port.write(exit_code as u32);
-    }
-}
-
-pub trait Testable {
-    fn run(&self) -> ();
-}
-
-impl<T> Testable for T
-where
-    T: Fn(),
-{
-    fn run(&self) {
-        // print name of the function
-        serial_print!("{}...\t", core::any::type_name::<T>());
-        self();
-        serial_println!("[ok]")
-    }
-}